pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trmnl-health"
version = "0.1.0"
description = "Pipeline that summarizes weight-loss metrics and pushes them to a TRMNL private plugin."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "typer>=0.12.3",
    "python-dotenv>=1.0.1",
]

[project.scripts]
trmnl-health = "trmnl_health.cli:app"
